Fix struct problem with long long int

+++ CMOD/src/TimeSpan.h
@@ -39,7 +39,7 @@ struct TimeSpan {
   float start;
   float duration;
   Ratio startEDU;
-  unsigned long long int startEDUAbsolute;
+  int startEDUAbsolute;
   Ratio durationEDU;
   
   TimeSpan() : start(0), duration(0), startEDU(0, 0), startEDUAbsolute(0), durationEDU(0, 0) {}
